feat: search fix web search
