+++ dev_gpt/options/generate/static_files/microservice/google_custom_search.py
@@ -1,30 +1,4 @@
 import os
-import openai
-
-
-openai.api_key = os.getenv("OPENAI_API_KEY")
-
-
-class GPT_3_5_Turbo:
-    def __init__(self, system_string: str = ''):
-        self.system = system_string
-
-    def __call__(self, prompt_string: str) -> str:
-        response = openai.ChatCompletion.create(
-            model="gpt-3.5-turbo",
-            messages=[{
-                "role": 'system',
-                "content": self.system
-            }, {
-                "role": 'user',
-                "content": prompt_string
-            }]
-        )
-        return response.choices[0]['message']['content']
-
-
-
-import os
 from typing import Optional
 
 import requests
@@ -52,4 +26,3 @@ def search_images(search_term, top_n):
 def search_web(search_term, top_n):
     response = google_search(search_term, search_type="web", top_n=top_n)
     return [item["snippet"] for item in response["items"]]
-
